Sets up main testing infrastructure.
Creates the tests/main.py and __init__.py files to set up the tests directory as a package.  The main.py script runs all tests
defined within the directory structure.  The reason for this is to allow for multiple versions of the api to be tested by creating
different directories for each version.