Create OneCodexResource as base class for all resources
This commit adds the OneCodexResource class as a superclass to all resources.
The purpose of this is to encapsulate URL generation for resources in a single place,
as well as centralize authentication object creation and other auxiliary functions.

Tests are also added for the newly refactored methods, which involve assigning a
resource name to every resource that inherits from OneCodexResource and
creating URLs with a resource name, and/or id, and/or action name.