Modify OneCodexAPIURLBuilder to be aggregated and not inherited
This commit changes the way OneCodexAPIURLBuilder is used.  Instead of resource classes
inheriting from it (which would require making assumptions about how URLs are built,
like requiring - or not - an ID), it is now meant to be instantiated.

There is a single instance per resource class (which means all instances of the same resource
access the same builder), which returns URLs for the API based on parameters.

The reason for this is to simplify URL construction and to not make resource classes have
to unnecessarily adhere to a set of assumptions, only to be able to generate valid API URLs.