Create the AnalysisDataAdapter class to handle analysis data.
This class works as an intermediary between the OneCodex API and the client library
by wrapping the data related to an analysis.  It allows access to formatted data
and should handle processing that may happen with analysis data.  As of now,
it converts the appropriate properties to their numeric types, and intelligently
returns None if they are not defined (instead of an error).